Move as much new code as possible to lua_privileges.cpp/h
