+++ rts/lib/lua/include/lauxlib.h
@@ -79,10 +79,6 @@ LUALIB_API int (luaL_loadfile) (lua_State *L, const char *filename);
 LUALIB_API int (luaL_loadbuffer) (lua_State *L, const char *buff, size_t sz,
                                   const char *name);
 LUALIB_API int (luaL_loadstring) (lua_State *L, const char *s);
-LUALIB_API int (luaL_loadfile_privileged) (lua_State *L, const char *filename);
-LUALIB_API int (luaL_loadbuffer_privileged) (lua_State *L, const char *buff, size_t sz,
-                                  const char *name, bool privileged);
-LUALIB_API int (luaL_loadstring_privileged) (lua_State *L, const char *s);
 
 LUALIB_API lua_State *(luaL_newstate) (void);
