+++ rts/lib/lua/src/lauxlib.cpp
@@ -22,6 +22,7 @@
 
 #include "lua.h"
 #include "lauxlib.h"
+#include "lua_privileges.h"
 
 #include "streflop_cond.h" // SPRING
 #include "System/BranchPrediction.h" // Recoil
@@ -636,26 +637,8 @@ LUALIB_API int luaL_loadbuffer (lua_State *L, const char *buff, size_t size,
 }
 
 
-LUALIB_API int luaL_loadbuffer_privileged (lua_State *L, const char *buff, size_t size,
-                                const char *name, bool privileged) {
-  LoadS ls;
-  ls.s = buff;
-  ls.size = size;
-  if (privileged)
-	  return lua_load_privileged(L, getS, &ls, name);
-  else
-	  return lua_load(L, getS, &ls, name);
-}
-
-
 LUALIB_API int (luaL_loadstring) (lua_State *L, const char *s) {
   return luaL_loadbuffer(L, s, strlen(s), s);
-}
-
-
-LUALIB_API int (luaL_loadstring_privileged) (lua_State *L, const char *s) {
-
-  return luaL_loadbuffer_privileged(L, s, strlen(s), s, true);
 }
 
 
@@ -686,5 +669,19 @@ LUALIB_API lua_State *luaL_newstate (void) {
   lua_State *L = lua_newstate(l_alloc, NULL);
   if (L) lua_atpanic(L, &panic);
   return L;
+}
+
+
+/* Recoil Privileged LoadBuffer */
+
+LUALIB_API int luaL_loadbuffer_privileged (lua_State *L, const char *buff, size_t size,
+                                const char *name, bool privileged) {
+  LoadS ls;
+  ls.s = buff;
+  ls.size = size;
+  if (privileged)
+         return lua_load_privileged(L, getS, &ls, name);
+  else
+         return lua_load(L, getS, &ls, name);
 }
